// src/app/models/partenaire.model.ts
export class Partenaire {
    actif:string;
    prenom: string;
    nom: string;
    id: string;
    paysResidance: string;
    mdp: string;
    adresse: string;
    logoUrl: string;
    nomPartenaire: string;
    couleur_primaire: string;
    adminId: string;
    //userId: string;

    constructor(){}
    
  }